fix: :bug: Improved usage of local storage
When on the select lap page and there is no passed filename in local storage, you are redirected to the page where you have to select a race and then the passed filename will be stored again. Removes the need for an extra api call. Plus now I don't have to guess which filename to use.